requirements.txt
requests
beautifulsoup4
lxml
pillow
imagehash
opencv-python-headless
numpy
pandas
pyarrow
playwright
tqdm
tenacity
